[docs] Corrected order of ControlPanel.test examples.
